+++ sdm/src/main/java/com/sap/cds/sdm/service/SDMService.java
@@ -33,7 +33,8 @@ public interface SDMService {
 
   public JSONObject getRepositoryInfo(SDMCredentials sdmCredentials) throws IOException;
 
-  public int deleteDocument(String cmisaction, String objectId, String user) throws IOException;
+  public int deleteDocument(String cmisaction, String objectId, String user, Boolean isSystemUser)
+      throws IOException;
 
   public void readDocument(
       String objectId, SDMCredentials sdmCredentials, AttachmentReadEventContext context)
@@ -80,4 +81,7 @@ public interface SDMService {
 
   public String getLinkUrl(String objectId, SDMCredentials sdmCredentials, boolean isSystemUser)
       throws IOException;
+
+  public byte[] readDocumentContent(
+      String objectId, SDMCredentials sdmCredentials, boolean isSystemUser) throws IOException;
 }