+++ sdm/src/test/java/unit/com/sap/cds/sdm/utilities/SDMUtilsTest.java
@@ -151,6 +151,16 @@ public class SDMUtilsTest {
   }
 
   @Test
+  public void testHasFileExtensionChanged() {
+    assertTrue(SDMUtils.hasFileExtensionChanged("sample.pdf", "sample.dmg"));
+    assertFalse(SDMUtils.hasFileExtensionChanged("sample.pdf", "renamed.pdf"));
+    assertFalse(SDMUtils.hasFileExtensionChanged(null, "file.txt"));
+    assertFalse(SDMUtils.hasFileExtensionChanged("file.txt", null));
+    assertFalse(SDMUtils.hasFileExtensionChanged("sample.pdf", "sample123"));
+    assertFalse(SDMUtils.hasFileExtensionChanged("sample", "sample123"));
+  }
+
+  @Test
   public void prepareSecondaryPropertiesTest_withFilenameKey() {
     Map<String, String> requestBody = new HashMap<>();
     Map<String, String> secondaryProperties = new HashMap<>();